Fixing view package name
