+++ src/com/jalasoft/search/view/AdvancedFieldsPanel.java
@@ -9,31 +9,26 @@
  * accordance with the terms of the license agreement you entered into
  * with Jalasoft.
  */
-package com.jalasoft.search.View;
+package com.jalasoft.search.view;
 import javax.swing.*;
 import java.awt.*;
 
 /**
  *
- This class implements the GUI for Search Menu Panel.
+ This class implements the GUI for advanced search panel.
  using components from swing library
  *
  * @version  1.0
  * @author Rodrigo Saul Cespedes Murillo
  */
-public class SearchMenu extends JPanel{
-    private JButton simpleButton;
-    private JButton advancedButton;
+public class AdvancedFieldsPanel extends JPanel {
 
     /**
      * Constructor initializes this Panel as grid layout also add components to this JPanel
      * */
-    public SearchMenu(){
-        this.setLayout(new GridLayout(4,1));
-        this.setBorder(BorderFactory.createTitledBorder("Search Menu"));
-        simpleButton = new JButton("Simple");
-        advancedButton = new JButton("Advanced");
-        this.add(simpleButton);
-        this.add(advancedButton);
+    public AdvancedFieldsPanel(){
+        this.setLayout(new GridLayout(1,2));
+        this.setBorder(BorderFactory.createTitledBorder("Advanced Search"));
     }
+
 }
